Fix upper bound for number of extensions with disconnected graphs
The code was calculating the number of extensions for a rightmost path
length that was not possible with disconnected graphs, because it was
always assuming that the rightmost path can cover all vertices.

This was resulting in the number of edges used by the rightmost path
exceeding the sum of degrees of the graph, which caused an unsigned int
underflow, leading to a very high upper bound.